+++ lib/config/options.rb
@@ -118,8 +118,6 @@ module Config
       result
     end
 
-    alias :to_h :to_hash
-
     def each(*args, &block)
       marshal_dump.each(*args, &block)
     end