+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-...
+### Bug fixes
+
+* Revert added alias to_h for to_hash ([#277](https://github.com/railsconfig/config/issues/277))
 
 ## 2.2.2