+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Bug fixes
 
-...
+* Revert added alias to_h for to_hash ([#277](https://github.com/railsconfig/config/issues/277))
 
 ### Changes